Improve SQLite History, Reorganize History
SQLite history backend now stores most datatypes in binary to preserve
data. SQL types are no longer relied on. The binary data is decoded
based on the UA Variant type that the value was originally stored with.
